Limpieza de codigo en componente Login
