+++ dental-landing/src/types/props/login.props.ts
@@ -1,0 +1,4 @@
+export interface LoginFormInputs {
+    email: string;
+    password: string;
+}
